fix(sdk): Remove deadlocks with Event Cache calling `Room` calling the Event Cache.
This patch fixes a bug where the `EventFocusedCacheState`
calling `Room::message` which calls `EventCache::save_events`
to insert an event. Problem: `EventFocusedCacheState` is behind
`StateLockWriteGuard`, and `EventCache::save_events` needs to acquire a
write lock over the state too. We have a deadlock then.

This is not the first time we see this bug: the Event Cache calling
a method that calls the Event Cache. The Event Cache must the sole
responsible to insert events in its storage, not another component.
Consequently, this patch removes the call to `EventCache::save_events`
in `Room::message`, along with 2 other places, making
`EventCache::save_events` unused, and a candidate for removal.
